project/requirements-dev.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alpaca-trade-api==3.2.0
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.1
apify_client==2.5.0
apify_shared==2.2.0
asttokens==3.0.1
async-generator==1.10
attrs==25.4.0
beautifulsoup4==4.14.3
blinker==1.9.0
Bottleneck==1.6.0
cachetools==7.0.5
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.6
curl_cffi==0.13.0
cycler==0.12.1
datasets==4.5.0
dateparser==1.2.2
decorator==5.2.1
deprecation==2.1.0
dill==0.4.0
distro==1.9.0
empyrical-reloaded==0.5.12
exchange_calendars==4.13.2
executing==2.2.1
Farama-Notifications==0.0.4
filelock==3.20.3
filetype==1.2.0
FinRL==0.3.7
fonttools==4.62.1
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.188.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.4
google-genai==1.70.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.80.0
grpcio-status==1.71.2
gymnasium==1.2.3
h11==0.16.0
hf-xet==1.2.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface_hub==1.3.4
idna==3.11
impit==0.12.0
importlib_metadata==8.7.1
ipython==9.8.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
korean-lunar-calendar==0.3.1
langchain==1.2.14
langchain-core==1.2.25
langchain-google-genai==4.2.1
langchain-openai==1.1.12
langdetect==1.0.9
langgraph==1.1.4
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.24
lxml==6.0.2
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.0.3
multidict==6.7.1
multiprocess==0.70.18
multitasking==0.0.12
mypy_extensions==1.1.0
narwhals==2.18.1
networkx==3.6.1
nltk==3.9.2
numpy==2.4.1
oauthlib==3.3.1
openai==2.30.0
orjson==3.11.8
ormsgpack==1.12.2
outcome==1.3.0.post0
packaging==26.0
pandas==2.2.3
parso==0.8.6
peewee==3.17.3
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.5.1
plotly==6.6.0
praw==7.8.1
prawcore==2.4.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.0
protobuf==5.29.6
psycopg2-binary==2.9.11
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pyfolio-reloaded==0.9.9
Pygments==2.20.0
pyluach==2.3.0
PyMuPDF==1.27.1
pyparsing==3.3.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.1
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
selenium==4.40.0
sentencepiece==0.2.1
setuptools==80.10.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.3
SQLAlchemy==2.0.49
stable_baselines3==2.8.0
stack-data==0.6.3
stockstats==0.6.8
streamlit==1.56.0
sympy==1.14.0
tabulate==0.10.0
tenacity==9.1.4
textblob==0.19.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
toml==0.10.2
toolz==1.1.0
torch==2.10.0
tornado==6.5.5
tqdm==4.67.1
traitlets==5.14.3
transformers==5.0.0
trio==0.32.0
trio-typing==0.10.0
trio-websocket==0.12.2
tweepy==4.16.0
typer-slim==0.21.1
types-certifi==2021.10.8.3
types-urllib3==1.26.25.14
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
update-checker==0.18.0
uritemplate==4.2.0
urllib3==1.26.20
uuid_utils==0.14.1
wcwidth==0.6.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==16.0
wrds==3.5.0
wsproto==1.3.2
xlsxwriter==3.2.9
xxhash==3.6.0
yarl==1.22.0
yfinance==1.1.0
zipp==3.23.0
zstandard==0.25.0
